refactor: All project structure
The whole project structure has been changed to get an architecture that
let to a future developer to make easy changes. Now if you want to
change JPA or implement other way to communicate with the API than HTTP
now it's possible.

Also good practices as Don't pass Entity models to the user and use DTOs
instead, apply some of solid principles and more things

It's a big change and possibly a lot of features are broken
in this commit (in special auth stuff), but it's something necessary to
have a good base for a next developer.